feat: task_id贯穿检测任务链路-后端任务落库与多帧映射

+++ RunDataPanel.cpp
@@ -138,7 +138,7 @@ void RunDataPanel::clearMetrics()
     }
 }
 
-void RunDataPanel::setMillingTask(int taskId, int pathTotal, bool calibrationApplied)
+void RunDataPanel::setMillingTask(quint64 taskId, int pathTotal, bool calibrationApplied)
 {
     if (m_taskIdLabel) {
         m_taskIdLabel->setText(QString("任务 #%1 · 共 %2 条")
